ignored .idea, added concat, separated directive into separate file
