+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 .DS_Store
 _site
 /assets/admin
+.idea/
