+++ assets/js/main.js
@@ -1,22 +1,6 @@
   angular.module('personalWebsite',[])
-  .controller('pwController', ['$scope', '$http', PwController])
-  .directive('animateScroll', function(){
-    return {
-      restrict: 'A',
-      scope: {
-        scrollTo: '@'
-      },
-      link: function(scope, $elm, attr){
-        $elm.on('click', function(){
-          console.log('click');
-          $('html,body').animate({scrollTop: $('#' + scope.scrollTo).offset().top - 40 +'px'}, 450);
-        });
-      }
-    };
-  });
-
-
-
+  .controller('pwController', ['$scope', '$http', PwController]);
+  
 function PwController ($scope, $http){
   $scope.test = "Designer";
   $scope.showNav = false;
