+++ assets/js/min/main.js
@@ -1,1 +1,1 @@
-function PwController(o,n){o.test="Designer",o.showNav=!1,n.get("personal_website.json").then(function(n){o.workItems=n.data,o.workTotal=o.workItems.length})}angular.module("personalWebsite",[]).controller("pwController",["$scope","$http",PwController]).directive("animateScroll",function(){return{restrict:"A",scope:{scrollTo:"@"},link:function(o,n,t){n.on("click",function(){console.log("click"),$("html,body").animate({scrollTop:$("#"+o.scrollTo).offset().top-40+"px"},450)})}}}),$(document).ready(function(){$(window).scroll(function(){$(window).scrollTop()<=100&&$(".main-nav").css({background:"rgba(255,255,255,"+Math.floor($(this).scrollTop())/100+")"})})});
+function PwController(o,n){o.test="Designer",o.showNav=!1,n.get("personal_website.json").then(function(n){o.workItems=n.data,o.workTotal=o.workItems.length})}angular.module("personalWebsite",[]).controller("pwController",["$scope","$http",PwController]),$(document).ready(function(){$(window).scroll(function(){$(window).scrollTop()<=100&&$(".main-nav").css({background:"rgba(255,255,255,"+Math.floor($(this).scrollTop())/100+")"})})});
