+++ assets/js/min/smooth-scroll-directive.js
@@ -1,0 +1,1 @@
+angular.module("personalWebsite").directive("animateScroll",function(){return{restrict:"A",scope:{scrollTo:"@"},link:function(o,l,c){l.on("click",function(){console.log("click"),$("html,body").animate({scrollTop:$("#"+o.scrollTo).offset().top-40+"px"},450)})}}});
