+++ gulpfile.js
@@ -3,6 +3,7 @@ var sass         = require('gulp-sass');
 var watch        = require('gulp-watch');
 var uglify       = require('gulp-uglify');
 var browserSync  = require('browser-sync');
+var concat       = require('gulp-concat');
 
 gulp.task('sass', function () {
   return gulp.src('assets/css/main.sass')
@@ -11,11 +12,17 @@ gulp.task('sass', function () {
     .pipe(browserSync.stream());
 });
 
-gulp.task('js', function() {
+gulp.task('minjs', function() {
   return gulp.src('assets/js/*.js')
     .pipe(uglify())
-    .pipe(gulp.dest('assets/js/min'))
+    .pipe(gulp.dest('assets/js/min/'))
     .pipe(browserSync.stream());
+});
+
+gulp.task('concatjs', ['minjs'], function() {
+  return gulp.src('assets/js/min/*.js')
+      .pipe(concat('allJs.min.js'))
+      .pipe(gulp.dest('assets/js/concat/'));
 });
 
 gulp.task('html', function(){
@@ -35,3 +42,6 @@ gulp.task('browserSync', function() {
 });
 
 gulp.task('default', ['browserSync']);
+
+
+gulp.task('js', ['minjs', 'concatjs']);
