+++ index.html
@@ -39,11 +39,8 @@
     </div>
   </div>
 
-
-
-
 <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.0.0-beta1/jquery.min.js"></script>
 <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.4.5/angular.min.js"></script>
-<script src="assets/js/min/main.js"></script>
+  <script src="assets/js/concat/allJs.min.js"></script>
 </body>
 </html>
