+++ package.json
@@ -15,5 +15,8 @@
     "gulp-sass": "^2.3.1",
     "gulp-uglify": "^1.5.3",
     "gulp-watch": "^4.3.5"
+  },
+  "dependencies": {
+    "gulp-concat": "^2.6.0"
   }
 }
